requirements.txt
colored-traceback==0.3.0
colorlog==6.6.0
cycler==0.10.0
decorator==4.4.2
distro==1.4.0
easydict==1.9
faiss-gpu==1.6.3
h5py==2.10.0
imagesize==1.3.0
importlib-metadata 
importlib-resources==5.4.0
matplotlib==3.3.4
mrcfile==1.1.2
nptyping==1.4.4
numpy==1.15.4
oauthlib==3.0.1
opencv-python==4.5.5.64
overrides==6.1.0
pandas==1.1.5
parso==0.7.0
Pillow==8.4.0
pipx==0.15.1.3
progress==1.5
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
python-dateutil==2.8.1
pytorch-metric-learning==1.3.0
pytz==2020.1
PyYAML==5.4.1
scikit-learn==0.24.2
scipy==1.1.0
seaborn 
tensorboard
tensorflow-tensorboard==1.5.1
toolz==0.10.0
tqdm==4.64.0
